+++ packages/reactabular-dnd/CHANGELOG.md
@@ -2,7 +2,7 @@
 ==================
 
   * Feature - Expose `dnd.draggableRow`. This allows you to use row dragging with virtualization. Example: `body.row = dnd.draggableRow(Virtualized.BodyRow)`.
-  * Feature - `dnd.draggableRow` will hook into row level (`onRow({ rowId })`) `onMoveStart`, `onMove({ sourceRowId, targetRowId })`, and `onMoveEnd({ rowId })` instead of just `onMove`.
+  * Feature - `dnd.draggableRow` hooks into row level (`onRow({ rowId })`) props `onCanMove({ rowId })`, `onMoveStart({ rowId })`, `onMove({ sourceRowId, targetRowId })`, and `onMoveEnd({ rowId })` instead of just `onMove`.
 
 7.0.0 / 2016-11-03
 ==================