+++ packages/reactabular/CHANGELOG.md
@@ -4,7 +4,7 @@
 ## reactabular-dnd
 
   * Feature - Expose `dnd.draggableRow`. This allows you to use row dragging with virtualization. Example: `body.row = dnd.draggableRow(Virtualized.BodyRow)`.
-  * Feature - `dnd.draggableRow` will hook into row level (`onRow({ rowId })`) `onMoveStart`, `onMove({ sourceRowId, targetRowId })`, and `onMoveEnd({ rowId })` instead of just `onMove`.
+  * Feature - `dnd.draggableRow` hooks into row level (`onRow({ rowId })`) props `onCanMove({ rowId })`, `onMoveStart({ rowId })`, `onMove({ sourceRowId, targetRowId })`, and `onMoveEnd({ rowId })` instead of just `onMove`.
 
 ## reactabular-easy